backend/services/rag/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.1
google-generativeai==0.3.2
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.3 